+++ lib/gnutls_ui.c
@@ -288,7 +288,7 @@ int gnutls_certificate_client_get_request_status(gnutls_session session)
   *
   **/
 int gnutls_fingerprint(gnutls_digest_algorithm algo, const gnutls_datum* data, 
-	unsigned char* result, int* result_size)
+	unsigned char* result, size_t* result_size)
 {
 	GNUTLS_HASH_HANDLE td;
 	int hash_len = _gnutls_hash_get_algo_len(algo);
