+++ lib/x509/dn.h
@@ -15,11 +15,11 @@
 
 int _gnutls_x509_parse_dn(ASN1_TYPE asn1_struct, 
 	const char* asn1_rdn_name, char *buf,
-	int* sizeof_buf);
+	size_t* sizeof_buf);
 
 int _gnutls_x509_parse_dn_oid(ASN1_TYPE asn1_struct, 
 	const char* asn1_rdn_name, const char* oid, int indx, char *buf,
-	int* sizeof_buf);
+	size_t* sizeof_buf);
 
 int _gnutls_x509_set_dn_oid(ASN1_TYPE asn1_struct, 
 	const char* asn1_rdn_name, const char* oid, const char *name,
